Address review feedback: 401 guards, shared waitFor helper, header cleanup
- Add 401 negative-auth tests to both describe blocks (TEST-13 blocker).
- Extract the inline waitFor polling helper into apps/web/tests/helpers/wait-for.ts; remove the duplicate from workspace-create-via.test.ts.
- Rewrite the file header so rationale is stated inline rather than by citing rule IDs / criteria-file paths.
